move files around, add tests for utils
