+++ lib/setup/hook-script.js
