+++ lib/setup/hook-script.test.js
@@ -13,7 +13,7 @@ const read = (dir, filePath) => fs.readFileSync(path.join(dir, filePath), 'utf-8
 
 const exists = (dir, filePath) => fs.existsSync(path.join(dir, filePath));
 
-describe('lib/setup/utils/hook-script', () => {
+describe('lib/setup/hook-script', () => {
   let dir;
   beforeEach(() => { dir = tempy.directory(); });
   afterEach(() => rimraf.sync(dir));
