+++ lib/setup/uninstall.js
@@ -1,9 +1,8 @@
 'use strict';
 
 const hooks = require('./hooks');
-const findHooksDir = require('./utils/find-hooks-dir');
-const hookScript = require('./utils/hook-script');
-const isSetupRequired = require('./utils/is-setup-required');
+const hookScript = require('./hook-script');
+const { findHooksDir, isSetupRequired } = require('./utils');
 
 const debug = require('debug')('gint:setup');
